# openapi.yml
openapi: 3.0.3
info:
  title: Testimonies.world API
  version: 1.0.0
  description: Casework platform for documenting human rights cases. Sensitive endpoints
    (contacts, sensitive media) require authentication.
paths:
...
  /api/testimonials/:
    get:
      operationId: testimonials_list
      description: |-
        Read + write API for testimonials.

        Public list / detail read uses `TestimonialPublicSerializer` for
        everyone — masking is the same regardless of role on the public
        endpoint. Internal endpoints (the workflow audit info) require
        Advocate+ via the write-side permission class.

        Permissions layered per action:
          list / retrieve             public (public serializer)
          create                      CanSubmitTestimonial
          update / partial_update     CanSubmitTestimonial + owner OR Advocate
          destroy                     CanPublishTestimonial (rare — archives
                                      are preferred over deletes)
          submit                      CanSubmitTestimonial
          approve / reject            CanReviewTestimonial
          publish / archive           CanPublishTestimonial
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: string
          enum:
          - approved
          - archived
          - draft
          - published
          - rejected
          - under_review
        description: |-
          Filter by workflow status. Role-scoped:
           - `?status=draft` narrows to the requester's OWN drafts for any authenticated user ("My drafts" semantic).
           - `?status=under_review`, `?status=rejected`, and `?status=published` widen for Advocate+ so the Review queue / Rejected tab / Published list surface actually shows the whole org.
           - Volunteers on any non-DRAFT status get a defensive narrow to their own rows.
           - Anonymous callers get PUBLISHED-only regardless of the value.
          Unknown values fall through to the role-based default (no narrowing).
      tags:
      - testimonials
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTestimonialPublicList'
          description: ''
    post:
      operationId: testimonials_create
      description: |-
        Read + write API for testimonials.

        Public list / detail read uses `TestimonialPublicSerializer` for
        everyone — masking is the same regardless of role on the public
        endpoint. Internal endpoints (the workflow audit info) require
        Advocate+ via the write-side permission class.

        Permissions layered per action:
          list / retrieve             public (public serializer)
          create                      CanSubmitTestimonial
          update / partial_update     CanSubmitTestimonial + owner OR Advocate
          destroy                     CanPublishTestimonial (rare — archives
                                      are preferred over deletes)
          submit                      CanSubmitTestimonial
          approve / reject            CanReviewTestimonial
          publish / archive           CanPublishTestimonial
      tags:
      - testimonials
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestimonialWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TestimonialWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TestimonialWriteRequest'
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestimonialWrite'
          description: ''
...
components:
  schemas:
...
    BlankEnum:
      enum:
      - ''
...
    PaginatedTestimonialPublicList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/TestimonialPublic'
...
    TestimonialLocationVisibilityEnum:
      enum:
      - public_precise
      - public_region
      - public_country
      - hidden
      type: string
      description: |-
        * `public_precise` - Public — precise
        * `public_region` - Public — region only
        * `public_country` - Public — country only
        * `hidden` - Hidden
    TestimonialPublic:
      type: object
      description: |-
        Public-facing payload — masks source + location by visibility flags.

        Server-side masking is the security boundary (SYSTEM_RULES §10).
        The frontend never sees redacted fields in the first place.

        The `source_visible` boolean captures whether *any* source
        descriptor is shown publicly (public_named / public_anonymous).
        Hidden rows surface `null`.
      properties:
        id:
          type: integer
          readOnly: true
        slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        title:
          type: string
          readOnly: true
        language:
          type: string
          readOnly: true
        country:
          type: string
          readOnly: true
        region:
          type: string
          readOnly: true
        public_location_display:
          type: string
          readOnly: true
        incident_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        incident_types:
          readOnly: true
        summary:
          type: string
          readOnly: true
        narrative:
          type: string
          readOnly: true
        outcome:
          type: string
          readOnly: true
        verification_level:
          allOf:
          - $ref: '#/components/schemas/VerificationLevelEnum'
          readOnly: true
        source_visible:
          type: boolean
          readOnly: true
        public_source_label:
          type: string
          readOnly: true
        family_protected:
          type: boolean
          readOnly: true
        contact_protected:
          type: boolean
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/TestimonialStatusEnum'
          readOnly: true
        published_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TestimonialTag'
          readOnly: true
        translation_group:
          type: string
          format: uuid
          readOnly: true
        linked_person_id:
          type: integer
          readOnly: true
        linked_report_id:
          type: integer
          readOnly: true
        schema_version:
          type: integer
          readOnly: true
      required:
      - contact_protected
      - country
      - family_protected
      - id
      - incident_date
      - incident_types
      - language
      - linked_person_id
      - linked_report_id
      - narrative
      - outcome
      - public_location_display
      - public_source_label
      - published_at
      - region
      - schema_version
      - slug
      - source_visible
      - status
      - summary
      - tags
      - title
      - translation_group
      - verification_level
    TestimonialSourceVisibilityEnum:
      enum:
      - public_named
      - public_anonymous
      - hidden
      type: string
      description: |-
        * `public_named` - Public — named
        * `public_anonymous` - Public — anonymous
        * `hidden` - Hidden
    TestimonialStatusEnum:
      enum:
      - draft
      - under_review
      - approved
      - published
      - rejected
      - archived
      type: string
      description: |-
        * `draft` - Draft
        * `under_review` - Under review
        * `approved` - Approved
        * `published` - Published
        * `rejected` - Rejected
        * `archived` - Archived
    TestimonialTag:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 64
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
          maxLength: 255
      required:
      - id
      - name
...
    TestimonialWrite:
      type: object
      description: |-
        Create / update shape for any role (Volunteer can create drafts).

        Server-controlled fields are read-only and silently dropped on
        input (SYSTEM_RULES §5):
          - status + all `*_by` + `*_at` workflow fields: only mutated
            via the transition actions, never via direct PATCH.
          - schema_version + is_exported: server-pinned.
          - created_by / created_at / updated_at: server-pinned.
          - source_encrypted + precise_location_encrypted: never
            serialized. Caller uses `set_source` / `set_precise_location`
            accessors (exposed via dedicated action endpoints when
            permission is held).
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 255
        slug:
          type: string
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        language:
          type: string
          maxLength: 10
        country:
          type: string
          maxLength: 100
        region:
          type: string
          maxLength: 255
        incident_date:
          type: string
          format: date
          nullable: true
        incident_types: {}
        summary:
          type: string
        narrative:
          type: string
        outcome:
          type: string
        source_visibility:
          $ref: '#/components/schemas/TestimonialSourceVisibilityEnum'
        public_source_label:
          type: string
          maxLength: 255
        location_visibility:
          $ref: '#/components/schemas/TestimonialLocationVisibilityEnum'
        public_location_display:
          type: string
          maxLength: 500
        family_protected:
          type: boolean
        contact_protected:
          type: boolean
        verification_level:
          oneOf:
          - $ref: '#/components/schemas/VerificationLevelEnum'
          - $ref: '#/components/schemas/BlankEnum'
        person:
          type: integer
          nullable: true
          description: Underlying Person this testimonial is built from.
        report:
          type: integer
          nullable: true
          description: Optional Report-level link (one event, one testimonial).
        tags:
          type: array
          items:
            type: integer
      required:
      - id
    TestimonialWriteRequest:
      type: object
      description: |-
        Create / update shape for any role (Volunteer can create drafts).

        Server-controlled fields are read-only and silently dropped on
        input (SYSTEM_RULES §5):
          - status + all `*_by` + `*_at` workflow fields: only mutated
            via the transition actions, never via direct PATCH.
          - schema_version + is_exported: server-pinned.
          - created_by / created_at / updated_at: server-pinned.
          - source_encrypted + precise_location_encrypted: never
            serialized. Caller uses `set_source` / `set_precise_location`
            accessors (exposed via dedicated action endpoints when
            permission is held).
      properties:
        title:
          type: string
          maxLength: 255
        slug:
          type: string
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        language:
          type: string
          minLength: 1
          maxLength: 10
        country:
          type: string
          maxLength: 100
        region:
          type: string
          maxLength: 255
        incident_date:
          type: string
          format: date
          nullable: true
        incident_types: {}
        summary:
          type: string
        narrative:
          type: string
        outcome:
          type: string
        source_visibility:
          $ref: '#/components/schemas/TestimonialSourceVisibilityEnum'
        public_source_label:
          type: string
          maxLength: 255
        location_visibility:
          $ref: '#/components/schemas/TestimonialLocationVisibilityEnum'
        public_location_display:
          type: string
          maxLength: 500
        family_protected:
          type: boolean
        contact_protected:
          type: boolean
        verification_level:
          oneOf:
          - $ref: '#/components/schemas/VerificationLevelEnum'
          - $ref: '#/components/schemas/BlankEnum'
        person:
          type: integer
          nullable: true
          description: Underlying Person this testimonial is built from.
        report:
          type: integer
          nullable: true
          description: Optional Report-level link (one event, one testimonial).
        tags:
          type: array
          items:
            type: integer
...
    VerificationLevelEnum:
      enum:
      - level_1_reported
      - level_2_partially_verified
      - level_3_corroborated
      - level_4_documented
      type: string
      description: |-
        * `level_1_reported` - Level 1 — Reported
        * `level_2_partially_verified` - Level 2 — Partially verified
        * `level_3_corroborated` - Level 3 — Corroborated
        * `level_4_documented` - Level 4 — Documented
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
...